Function embed_resource::compile

pub fn compile(resource_file: &str, prefix: Option<&str>, out_dir: Option<&str>)

Compile the Windows resource file and update the cargo search path if we're on Windows.

prefix, if None, defaults to "$CARGO_PKG_NAME-manifest", this is the name you'll link to in main.rs.

out_dir, if None, defaults to $OUT_DIR, which you probably don't want to change.

On non-Windows this does nothing, on non-MSVC Windows, this chains windres with ar, but on MSVC Windows, this will try its hardest to find RC.EXE in Windows Kits and/or SDK directories (because someone thought not putting it in $PATH was a great idea).

Examples

In your build script, assuming the crate's name is "checksums":

extern crate embed_resource;

fn main() {
    // Compile file checksums.rc to be linkable as checksums-manifest in $OUT_DIR
    embed_resource::compile("checksums.rc", None, None);
}

If you want to link as chksum-rc:

extern crate embed_resource;

fn main() {
    // Compile file checksums.rc to be linkable as chksum-rc in $OUT_DIR
    embed_resource::compile("checksums.rc", Some("chksum-rc"), None);
}